+++ CHANGELOG.md
@@ -14,6 +14,7 @@ configured prefix accordingly.
 
 - Added config key `max-duration` (integer)
 - Added config key `max-duration-allow-unknown` (boolean)
+- Added function `util.compose()`
 
 ### Fixed